// src/models/orderInterface.ts
export interface IOrder{
    orderId: string
    orderDate: Date
    orderQuantity: number 
    productid:string
    productId: string
    productName: string
    coffeeShopId:string
    coffeeShopAddress: string
    customerId:string
    customerName: string
    customerPhone:string
    customerAddess: string
    paymentTypeId:string
    paymentTypeName: string
}